requirements.txt
sounddevice>=0.4.6
scipy>=1.11.0
openai>=1.0.0
pynput>=1.7.6
python-dotenv>=1.0.0
